vpnmanager: vpn-interfaces yang container has vpn-instance-name list
In order to handle internal VPN or BGPVPN in a two router case solution,
some yang changes is needed.
Config DS is being added a list of vpn-instance configuration for one
VPN interface. The need is to have two vpn instances configurable for one VPN.
This commit change only handles the compilation by adapting java code
to as to care about the first element of vpn-instance.

Change-Id: I87a4919899b0c702267977e5942ff8589f77de55
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>