vpnmanager: yang modify vpn-interfaces container. vpn-instance-name is a list
In order to handle internal VPN or BGPVPN in a two router case solution,
some yang changes is needed.
In a dual router case solution, one vpn Instance and two vpn Interfaces
will lead to create one structure for config DS and two structures for
oper DS.
Especially, config DS is being added a list of vpn-instance-name.
The need is to have two vpn instances configurable for one VPN.
This commit change only handles the compilation by adapting java code
to as to care about the first element of vpn-instance.

Change-Id: I87a4919899b0c702267977e5942ff8589f77de55
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>