vpnmanager: yang modify vpn-interfaces container. vpn-instance-name is a list
The need is to have two vpn instances configurable for one VPN
interface. That case can happen when having a VPN interface connected to
two BGPVPNs: one for inter-dc purposes, and the other one for external
connectivity.
A leaf list is done in the yang model.
The other modificatiosn consist in ensuring compilation through
vpnmanager. The first vpninstance of vpnInterface is retrieved.

Change-Id: I87a4919899b0c702267977e5942ff8589f77de55
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>