New (still empty) FibManagerComponentTest
This is a "skeleton" component test (see
https://wiki.opendaylight.org/view/BestPractices/Component_Tests) for
fibmanager.  Real functional tests should be filled in to this.

I would recommend that you merge this as-is already ASAP, instead of
waiting to fully completing it as part of this change, and then
separately raise new follow-up changes to (gradually) add real tests to
this skeleton.

Depends on https://git.opendaylight.org/gerrit/#/c/58085/ and on
removing the BundleContext in FibManagerImpl in (TBD).

Change-Id: I01779b5fe57dedf7eb246d583f0211bdd8b72682
Signed-off-by: Michael Vorburger <vorburger@redhat.com>