New (still empty) FibManagerComponentTest
This is a "skeleton" component test (see
https://wiki.opendaylight.org/view/BestPractices/Component_Tests) for
fibmanager.  Real functional tests should be filled in to this.

I would recommend that you merge this as-is already ASAP, instead of
waiting to fully completing it (as part of this change), and just later
separately raise new follow-up changes to (gradually) add real tests toh
this skeleton.

This change depends on:
* https://git.opendaylight.org/gerrit/#/c/58085/
* https://git.opendaylight.org/gerrit/#/c/58089/ 
* https://git.opendaylight.org/gerrit/#/c/58093/ 
* https://git.opendaylight.org/gerrit/#/c/58094/
* on removing the BundleContext in FibManagerImpl in (TBD).
* on either a fake or the real IElanService impl in (TBD)
* on either a fake or the real ItmRpcService impl in (TBD)

Change-Id: I01779b5fe57dedf7eb246d583f0211bdd8b72682
Signed-off-by: Michael Vorburger <vorburger@redhat.com>