Bug 8595 - DNAT traffic from DC gateway to FIP fails
Description :

In PDNAT_TABLE, we have a FIP mac match for performing FIP to internal IP
translation. But in L3_LFIB_TABLE, the destination mac has not been
updated and as a result, the reverse traffic for FIP were dropped.

Changes done to update packet with destination FIP mac as a result,
PDNAT_TABLE has specifc match and FIP to internal IP translation been done
properly.

Change-Id: I3fec733c7a8c049808a09828b8090daed12c83fc
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>