Bug 8441 - LB groups not deleted after extra route deletion
Issue :

When an extra route is completely removed from all the DPNs, the
loadbalancing groups are not updated properly.

Analysis:

There was a "NullpointerException" occuring in the
setupLoadbalancingGroup() method in the NexthopManager which caused the
issue. Also, there was an issue in the population of the adjacencylist in
the resolveadjacency() method which resulted in erratic programming of
the remote fib flows.

Fix:

Changes have been made to avoid the exception caused. Also, changes were
made in resolveadjacency() method to populate the adjacencylist
correctly. Further, there was an issue in removeorUpdateFibEntry()
method in FibUtil which caused multiple redundant calls to the update()
method of VrfEntryListener. It has been rectified too.

This patch also fixes the bug 8630 as well which resulted due to the
exceptions caused as mentioned above.

Change-Id: I32ebd1ebba6dc7eebdd358a2825990037d6ec432
Signed-off-by: Gobinath <gobinath@ericsson.com>