Update patch set 1
Patch Set 1:

Hi Vivek,

I did not include the changes to delete router interface flows for inactive dpns because I believe its not needed. Due to the change in NeutronPortChangeListener, the router interface FIB entry should get deleted first, and then the corresponding FIB entries should get deleted.

This, in turn, will ensure that the vpn-to-dpn will be rendered inactive only when the vpn interface op listener for the last interface on the dpn for the vpn will get notified. By, that time the vrfEntryListener notified to handle router interface FIB delete will already have read vpnInstanceOpData and it will contain all the vpn-to-dpn list in active state.

This will also ensure that there wont be any unnecessary "Modofied Node Does Not Exist" exception in the logs due to cleanupVpnOnDpn and deleteFibEntries trying to remove the flow at the same time.

Also, please note that this issue arises only for router interfaces and subnet-route entries, as discussed earlier. Now, due to the router interface handled properly in the NeutronPortChangeListener and subnet-route being inherently faster, the removal cycle for these entries will not be affected by the state of the vpn on a dpn.

Please let me know what you think.

Thanks,
Kiran

Patch-set: 1