Update patch set 1
Patch Set 1:

On a related note, I think we will have to think about what happens when a vpn is deleted, and simultaneously an external(bgp) fib entry is removed by bgpmanager. :)
This might potentially lead to stale table 20 and 21 flows, as cleanupDpnOnVpn and vrfEntry.remove() might get executed in parallel...

This I think can be handled in a separate review.

Patch-set: 1