VrfEntryListener refactoring using chain of responsibility pattern
This is a first review where BGP Route handling is separated into a new
handler class BgpRouteVrfEntryHandler.

All common methods are put into a new superclass i.e VrfEntryHandler. Each
of the subclasses like BgPRouteVrfEntryHandler will extend the superclass.
This is done to avoid circular depency on VrfEntryListener.

As part of chain of responsibility pattern , VrfEntryListener will pass on
the Vrf object to the first handler in a chain based on the RouteOrigin
type. Once the first handler completes processing the Vrf obj , it will
pass on the same to next handler.

In the current review BgpRouterVrfEntryHandler will process Routes of type
"RouteOrigin.BGP". BGP routes need not be processed by any other handler ,
hence next handler is not defined in the chain.

In the subsequent patches , when new handlers like
LocalRouteVrfEntryHandler, SubnetRouteVrfEntryHandler etc will have to
pass on the Vrf obj to next handler which will be
ImportExportVrfEntryHandler.

Change-Id:  I538085bfcda7733b5f3ba5ffe49d541619ac40ec
Signed-off-by: ehvkand <hanamantagoud.v.kandagal@ericsson.com>