Bug 8552 - [l3vpn]Stale flow entries not getting removed even when all the VM config are deleted
FIB handling was being interrupted by and NPE, which led to stale remote
table-21 flows. This is because the GroupEntityBuilder has been modified
with respect to the way it adds BucketListInfo.
It tries to addALL the elements from the input List of buckets now, and
hence a null param causes NPE. This has been fixed.

Change-Id: Iafe20e405101c294ce56babbaf5ff0057be55ae7
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>