Bug 8441 - LB groups not deleted after extra route deletion
Issue :

When an extra route is completely removed from all the DPNs, the
loadbalancing groups are not updated properly.

Analysis:

There was a "NullpointerException" occuring in the
setupLoadbalancingGroup() method in the NexthopManager which caused the
issue.

Fix:

Changes have been made to avoid the exception caused. Also, changes were
made in resolveadjacency() method to populate the adjacencylist
correctly.

This patch also fixes the bug 8630(CSIT Sporadic failures - arp learning suite - MIP not in /config/odl-fib:fibEntries/) as well which resulted due to the
exceptions caused as mentioned above.

Change-Id: Iaffd3916603f33810cfb7fdba2a86ba8826fc9e3
Signed-off-by: Gobinath <gobinath@ericsson.com>