Update patch set 2
Patch Set 2:

Vishal,
We discussed this in the hwvtep call, and I believe the consensus was to just go ahead and set this without checking schema versions, in general, as doing such checks would be clutter and not preferred going forward.

If you agree, please indicate such with a +1 (or +2). Otherwise, please let me know what you would like to see.

Thanks,
Vic

Patch-set: 2
Label: Verified=0