Update patch set 8
Patch Set 8:

(2 comments)

For PS3, Line 364: Is it really necessary to raise this exception ? 

From the specification for dualstack [0], Chapter "Configuration impact: 2. Limitations for subnetworks configurations":


2.2 Maximum numbers of different ethertype subnetworks associated to a one L3 BGPVPN instance is limited to two. If a network contains more than two different ethertype subnetworks, exception won't be raised, but this configuration isn't supported.

And from chapter "Proposed changes: 2. Support of two routers, attached to the same L3 BGPVPN: 2.1 Changes in neutronvpn"

For each created in dualstack network Neutron Port we take only the last received IPv4 address and the last received IPv6 address. So we also limit a length of subnets list, which could be attached to a L3 BGPVPN instance, to two elements.



[0] https://git.opendaylight.org/gerrit/#/c/54089/32/docs/specs/l3vpn-dual-stack-vm.rst,unified

Patch-set: 8
Label: Verified=0