vpninterface-manager: use adjacenciesOp in vpnInterfaceOpDataentry
This commit fixes some changes done in " introduce vpn-interface-op-data
context in oper DS". Instead of using the VpnInterface config DS, it
uses the AdjacenciesOp from vpnInterfaceOpDataEntry in operDS.

API is changed for the following:
o the vpnInterface identifier is replaced with vpnInterfaceOpDataentry
for addVpnInterfaceVpn and addNewAdjToVpnInterface and delAdjFromVpnInterface

o functions updateVpnInterfaceOnTepDelete and
updateVpnInterfaceOnTepAdd are reverted and use only
vpnInterfaceOpDataEntry as parameter.

o Two APIs have VPNInterface and identifier of VPNInterface replaced by
VPNInterfaceOpDataEntry and identifier of VPNInterfaceOpDataEntry
- withdrawAdjacenciesForVpnFromBgp
- advertiseAdjacenciesForVpnToBgp

Change-Id: I8afa91f3f4ff56d1b617bf4d0b7d16d4512f9fe4
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>