netvirt: various changes to support vpn-interface-oper DS
- natservice: misc nat changes while using vpn-interface-op-data-entry
On some cases where a VM interface is connected to two routers ( case
with dual stack IPs), then the router which has nat enabled is searched.
The changes parse the list of router per VPN interface, and return the
first router which has NAT enabled.
Obviously, that kind of setup may have only one router with NAT enabled,
at maximum.
On EVPN flow programmation, on removing floating IPs, the vpnInterfaces
operational is removed too.
On Interface state change, the vpnInterfaces will be updated, added or
removed.

- cloud-service-chain: update VpnToPseudoPortData with VpnInterfaces
Because VpnInterface can support more than one VPN, when a new VPN
interface is created or removed, the list of its VPN should be parsed
against the VpnToPseudoPortData contexts.

- fib-manager: update operational Adjacencies by using vpnInterfaceOpDataEntry
This commit deletes the adjacenciesOp entry related to a vrfEntry, in
vpnInstanceOpDataEntry. The list of vpnRouterIds() is parsed in config
DS vpnInstance, to check is there is a vpnInstanceOpDataEntry existing.
Then if present, the associated adjacenciesOp is deleted.

Change-Id: I0db2c7aecdc2d4c082e6ef37b3c62e38bc8edb2b
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Signed-off-by: Noel de Prandières <prandieres@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>