Remove optional in bgpmanager.xml
This shouldn't be needed. If it is, then there's another bigger problem,
which should be resolved, instead of cheating like this.. because if its
optional, it will be injected as null, but the code in
BgpConfigurationManager has no null checks for IVpnLinkService, so...
this could never have worked - isn't this kind of coded tested?! ;)

The https://git.opendaylight.org/gerrit/#/c/59421/ has a related issue,
but I chose to file two separate Gerrits.  Without c/59421, the
IVpnLinkService in BgpConfigurationManager is anyway always null...

Found chance related to a closer look in the context of
https://git.opendaylight.org/gerrit/#/c/59205/.

Change-Id: Ib82ab46825412572442c7b06b7f351f187ec4894
Signed-off-by: Michael Vorburger <vorburger@redhat.com>