Remove optional in elanmanager.xml
This shouldn't be needed. If it is, then there's another bigger problem,
which should be resolved, instead of cheating like this.. because if its
optional, it will be injected as null, but the code in EvpnUtils has no
null checks for IVpnManager and IBgpManager, nor does
ElanInterfaceManager for its INeutronVpnManager, so... if these services
were not available (i.e. truly optional), then this would just cause
NPEs.

Found chance related to a closer look in the context of
https://git.opendaylight.org/gerrit/#/c/59205/.

Change-Id: Icb52766bff22580d243d3d2f217f621f0d4c0184
Signed-off-by: Michael Vorburger <vorburger@redhat.com>