Remove optional in vpnmanager.xml
This shouldn't be needed. If it is, then there's another bigger problem,
which should be resolved, instead of cheating like this.. because if its
optional, it will be injected as null, but the code in
CentralizedSwitchChangeListener has no null checks for IVpnManager (and
the ArpMonitoringHandler/ArpMonitorStartTask doesn't even use the
INeutronVpnManager anyway), so... if these services were not available
(i.e. truly optional), then this would just cause NPEs.

Found chance related to a closer look in the context of
https://git.opendaylight.org/gerrit/#/c/59205/.

Change-Id: Ic34e47ae9dcaef0ac03a6772c9088dc2640a2209
Signed-off-by: Michael Vorburger <vorburger@redhat.com>