Bug 8743 : Fix NAT Service to safely use vpn-to-dpn list.
Description : Its been observed the vpn-to-dpn list been not updated
properly resulting in missing dpn-state with active state. As a result,
installation of default route rule been missing resulting in SNAT/DNAT
failure.

Changes are done to make use of vpn footprint service to update the
vpn-to-dpn list whenever floating-ip and external fixed ips are added
and deleted.

Change-Id: Ic55fc5364ef45fbbe7ce63a9feef51bf39332541
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>