vpninterface-manager: proceed VpnInterface state
If VPN instance is still not associated with router,
VpnInterfaceOpDataEntry will be created, but it will have state
INACTIVE, otherwise VpnInterfaceOpDataEntry will be created with state
ACTIVE. Subsequently, the VPN Interface state is maintained, thanks to
events coming from VpnInstanceListener. Upon those events, a list of
VPNInterfaceOpDataEntry having the same VPNName is retrieved. The VPN
Interface is changed accordingly.

Change-Id: Icbffcdbd3bb6ec24e6ddebfe4d56117f869039af
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Noel de Prandières <prandieres@6wind.com>
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>