Update patch set 14
Patch Set 14:

Hi Kiran,

I remember a discussion with Ashvin, asking for clarification about that field.
vpn-name was standing for the router name in that case.

The vpn-name change to vpn-router-ids is here to better reflect the signification of that value.

So I would say your proposal is incorrect. 
We want to keep to new value.
Please elaborate if you really need that change to be abandoned, and cc Ashvin/Vivek.

Also, please do not forget to revert the -1 or add +1 if you agree with us.
Thanks

Patch-set: 14
Reviewer: Gerrit User 6574 <6574@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0