Update patch set 18
Patch Set 18:

Hi Kiran,
about your remarks, what i did is the following:
- replaced Arrays.asList with Collections.singletonList
- set to abandon https://git.opendaylight.org/gerrit/#/c/59015/

for the other comment:
" use a variable multiple getVpnInstance.get(0) is called."
I consider that all subsequent commits remove the call to get(0).
So that all code should be compliant.
Is that solution ok ?
Or do you wish that variable to be introduced in that commit, however ?

Patch-set: 18
Label: Verified=0