Update patch set 18
Patch Set 18:

(2 comments)

Philippe,,

For now please move vpnInstanceNameList.get(0) to a separate method inside vpnHeler.java.  Use that method everywhere to retain the other files to compile.

After we merge in the refactoring reviews where the operational DS for vpn interface uses (vpninstancename + vpninterfacename) key, the method we introduce in this commit will be removed and all code will need to be changed to loop through all configured vpninstances for a vpnName.

Patch-set: 18