Bug 8789 - Designated DHCP DPN is missing.
When l2gw connection is creaated and first dpn interface creation
happens at about same time there is a race in dhcp designated listener
which thinks that it has to update mcast only if there are no dpns in
elan.

when an elan dpn is removed, the dpns tep ip should not be removed from
mcast mac if that tep ip is the designated dpn tep ip.

when l2gw is deleted clear its connections also.

Change-Id: I6a216cecc609ba75f10ebfc750e492348e447122
Signed-off-by: eaksahu <a.k.sahu@ericsson.com>