Bug 8346 - Conflicting modification for vpnNextHops.
The VpnInstanceListener used to wait for the vpn Interface Operational
data to clean up, before cleaning up data stores related to it.
There was a max wait time calculate, but some time interface op cleanup is
exceeding this wait time. As a result, the vpn instance related cleanup is
done before or at the same time as the vpn interface op cleanup,
resulting in these exceptions.
This fix remove the timed wait for the vpn instance, and cleans up the vpn
data based on notifications fired when both a vpn instance delete is
orchestrated and its footprint is cleared

Change-Id: I54d6369f30aa6d2a0e98df56b80cfcd6b4295978
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>
Signed-off-by: Sam Hague <shague@redhat.com>