Use manual setters instead @Immutable in IdentifiedAceBuilder.xtend
The mix of Immutables.org with Xtend.org seemed cool when I originally
came up with this 1 year ago, but after having run into issues with
in-IDE compilation order in Eclipse EVERY TIME I want to touch something
in aclservice, this got enough on my nerves to finally simplify this.

The issue basically is that the Xtend IDE and Eclipse APT processor seem
to struggle about the correct order in which Xtend active annotations
and APT processors should run (although it did, always, work, reliably,
in the mvn CLI build); Eclipse IDE could never get this right, and there
was too often red ink errors for (supposedly) missing
ImmutableIdentifiedAceBuilder... so basically this mix of Xtend with
Immutables.org probably wasn't a good idea.

The Xtend active annotations @Accessor (as well as @Data) really do
something quite similar to Immutables.org (yet let us keep the "pretty"
object literal initialization syntax of Xtend); however they generate
"real" setters, which do not "return this;", so instead it's simplest to
just hand write those few setters, for now.  Later if this pattern is
used more widely, we could have a new Xtend active annotation, such as
@Builder.

This also, and strictly speaking kind of unrelated, changes the double
indirection of Builder -> Immutable -> DataObject to a simpler to
understand single level, without the intermediate Immutable (which makes
these Builders used to compose objects more easily in tests not thread
safe, which is perfectly fine, actually).

dito for IdentifiedInterfaceWithAclBuilder

Change-Id: Ia70f5f62fa9d3d71da0521cd22808907c97a4619
Signed-off-by: Michael Vorburger <vorburger@redhat.com>