Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Very minor pt.  Other than that, I'm good with this, and will +1 after. 

BTW: On inevitable back-port to stable/* which some may be tempted to immediately do, personally I would keep this on master only - just on the off chance that we missed an impact and this could lead to NPE (iwhich it should not, of course) - this is the sort of code clean up action which I think is "nice" to do, but would never back-patch, just out of extra caution.

Patch-set: 1