vpnmanager: various support for vpn-interface-op in operDS
Misc changes in vpn-manager related to the support of two router setup:
- tunnelinterface state changes
This commit enhances API changes for the following:
o updateVpnInterfaceOnTepDelete
o updateVpnInterfaceOnTepDelete
Instead of looking for a VPNInterface in operDS and refreshing context,
the VPNInterfaceOpDataEntries are looked up. Contexts are refreshed
accordingly.
- vpn-subnet: update packet processing with correct router-id
The metadata retrieved from the packet is compared against the router
used in the VPN interface. Action is done accordingly.
Also, upon interface going down, a check is done on the retrieving
datapath ID and performing subnet propagation, provided that
vpninterface-op-data-entry is present.

- vpn-util: check ARP learning against VPN interface
When receiving an ARP response on a VPN interface, the metadata is
checked against the list of routers in that VPN interface.

- vpnmanager-interface-op-data: adapt vpnInterfaceOpDataEntry listener
This change handles the update and deletion of operDS vpnInterfaceOpDataEntry
 entries. It replaces the old code that was expecting vpnInterface entries.

- vpnmanager-shell: dump the list of vpnInterfaceOpDataEntry()
This lists the whole list of vpnInterfaceOpDataEntry() previously
created when a vpnInterface was created in config DS.

Change-Id: Ia0a8845464912a56a16e87649c2adaf74c71d5de
Signed-off-by: Noel de Prandieres <prandieres@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>