Dualstack support for L3VPN - single router Dual stack
According to dualstack specification [0], this commit introduces changes
related to single router dual stack case.

* NeutronPortChangeListener: Create only one VPN interface instance
  for one neutron port. This instance can include a list of IPv4 and/or
  IPv6 subnetMaps, where addresses where allocated for spawned port.

* NeutronvpnManager: in createVpnInterface we no longer use routerId as
  argument, instead we will iterate over all port IP addresses to create
  according adjacencies objects and than we pass composed adjacencies
  list to writeVpnInterfaceToDs method.

A VPN instance node has ipv4 and ipv6 ready attributes. The modification
does the following: according to the nature of the router with which a
VPN is attached, the ipv4 attribute is instanciated or the ipv6
attribute is instanciated, or both attributes are instanciated. On VPN
instance update, parse the ipv4 or ipv6 family attributes of the VPN
instance and update the RDs accordingly.

vpnmanager: update VpnInstance with IpFamily

When subnet is attached or withdrawed to/from VpnInstance, update
VpnInstance IpFamily attributes in dependency of added or removed subnet
IpFamily type.

Change-Id: Ie871496b55b05453ab6f16277fa3ca9cf9f2adaf
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>