Dualstack support for L3VPN - single router Dual stack - test
According to dualstack specification [0], this commit introduces changes
related to "one router solution" case.

* NeutronPortChangeListener: Create only one VPN interface instance
  for one neutron port. This instance can include a list of IPv4 and/or
  IPv6 subnetMaps, where addresses where allocated for spawned port.

* NeutronvpnManager: in createVpnInterface we no longer use routerId as
  argument, instead we will iterate over all port IP addresses to create
  according adjacencies objects and than, we pass composed adjacencies
  list to writeVpnInterfaceToDs method.

neutronvpn: create ipv4 or ipv6 context

Add ipv4Family and ipv6Family attributes to VPN instance. Add
IpVersionChoice class, which allows to set values for these attributes
in dependency of Ethertype of added/removed to VPN subnet.

bgpmanager: change API of bgpmanager to add VRF IPv4 or IPv6

This change adds a change in IBgpManager API, since it adds afi and safi
parameters to VRF creation and deletion. This information is propagated
to BGP so as to create and delete VPN context tables in function of the
AFI and SAFI settings.
Instead of passing a layerType value, an AddressFamily type is used.
AddressFamily is mapped to internal afi/safi values.
The values are appended or removed to the VRF objects of BGP Manager.
The VRF object is then written to the config DS.
Subsequently, the Quagga will be called upon the various changes in the VRF object.
If the object is added or updated, the Quagga will be called with the appropriate
afi/safi values. If the object is removed, all the afi/safi values are removed from
the Quagga. If the object is added, the afi/safi values are added to the quagga.

bgp-manager: update thrift model

addVrf() and delVrf() have 2 new parameters that are needed.
The modification consists in adapting the necessary thrift generated
files changes. On top of that, the call appropriate to BgpRoute.addVrf
and BgpRoute.delVrf is done with correct parameters.

Bug-9134: avoid returning exception when checking for a non IPv4 prefix

In order to not having an exception returned, the Nwutils library is not
used to check for an IPv4 address.

vpnmanager: update vpn instance families based on subnetmap updates only

Upon a change in configuration, VPN instance from config DS is updated.
The change is due to a subnetmap moving to or from an other vpnid. If the old
VPN was a BGPVPN, then the list of subnetmaps will be parsed to check if
an IP family is missing or not. Reversely, the introduction of a new
subnet refreshes the IP families of the VPN instance.
According to the information, the VPN instance is updapted.
Consequently, if BGPVPN is configured, then the associated BGP VRFs are
updated accordingly.
Also, the vpnmanager shell command is modified accordingly with correct
AFConfig. Because a VPN instance can be either IPv4 or IPv6 or both,
then a check should be done against the good structure to parse.

Bug-9153:  cloud service chain / compilation error

The yang model date revision has been changed to comply with the
compilation.

Change-Id: Ifd2a6998f238678a8a3b35a135e7fd26c9e9d425
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>