Dualstack support for L3VPN Single router Dual stack - nitrogen
According to dualstack specification [0], this commit introduces changes
related to "one router solution" case.

* NeutronPortChangeListener: Create only one VPN interface instance
  for one neutron port. This instance can include a list of IPv4 and/or
  IPv6 subnetMaps, where addresses where allocated for spawned port.

* NeutronvpnManager: in createVpnInterface we no longer use routerId as
  argument, instead we will iterate over all port IP addresses to create
  according adjacencies objects and than, we pass composed adjacencies
  list to writeVpnInterfaceToDs method.

neutronvpn: create ipv4 or ipv6 context

Add ipv4Family and ipv6Family attributes to VPN instance. Add
IpVersionChoice class, which allows to set values for these attributes
in dependency of Ethertype of added/removed to VPN subnet.

neutronvpn: update VpnInstance with IpFamily type

When subnet is attached or withdrawed to/from VpnInstance, update
VpnInstance Ipv4Family, Ipv6Family  attributes in dependency of added or
removed subnet's Ethertype.

bgpmanager: change API of bgpmanager to add VRF IPv4 or IPv6

This change adds a change in IBgpManager API, since it adds afi and safi
parameters to VRF creation and deletion. This information is propagated
to BGP so as to create and delete VPN context tables in function of the
AFI and SAFI settings.
Instead of passing a layerType value, an AddressFamily type is used.
AddressFamily is mapped to internal afi/safi values.
The values are appended or removed to the VRF objects of BGP Manager.
The VRF object is then written to the config DS.
Subsequently, the Quagga will be called upon the various changes in the VRF object.
If the object is added or updated, the Quagga will be called with the appropriate
afi/safi values. If the object is removed, all the afi/safi values are removed from
the Quagga. If the object is added, the afi/safi values are added to the quagga.

vpnmanager: create bgpmanager according to vpninstance type

bgp-manager: update thrift model

addVrf() and delVrf() have 2 new parameters that are needed.
The modification consists in adapting the necessary thrift generated
files changes. On top of that, the call appropriate to BgpRoute.addVrf
and BgpRoute.delVrf is done with correct parameters.

vpnmanager-shell: associate RD with correct AfConfig

Because a VPN instance ca be either IPv4 or IPv6, or both, then a check
should be done against the good structure to parse.

Change-Id: I6ac7dc61470c45e10c82e4faf2992a5572016e07
Signed-off-by: Noel de Prandieres <prandieres@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>