Update patch set 1
Patch Set 1:

Sam, Suneel
Thanks for your comments.

We want to fix this bug as we think it's fatal.
However, we are still not very familiar with ODL and HWVTEP,
so we can fix this bug in the partial range, like the patches we pushed,
but it's difficult for us to fix this bug with the way you suggested, 
as its influence range is wide.
So we wonder if you can help to fix?

Patch-set: 1
Reviewer: Gerrit User 7423 <7423@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0