Update patch set 2
Patch Set 2:

(1 comment)

It should not be null. If null aclservice cannot program rules correctly. The reason seems to be the timings issues, where the port update is called even before lport tag is generated. This patch does not address the root cause(seems to be not in aclservice).

Patch-set: 2
Label: Verified=0