Update patch set 18
Patch Set 18:

thanks Abhinav for your remarks.
we fully understand your concern about the added tracing that is displayed by default.
the fact is right now, this commit completely breaks the CSIT. so we want to understand why it breaks by adding some usefule traces.

once we know what needs to be done, we will revert the traces.
thanks,

Patch-set: 18