Bug 8889 - Moved trailing comments onto previous empty line.
This rule verifies that single-line comments are not located at the end of a line of code.
The main idea behind this rule is that in order to be really readable, trailing comments would have to be properly written and formatted
(correct alignment, no interference with the visual structure of the code, not too long to be visible) but most often,
automatic code formatters would not handle this correctly: the code would end up less readable.
Comments are far better placed on the previous empty line of code, where they will always be visible and properly formatted.

Change-Id: I8df4d06db88a80ed2f92d1d86bde6fd0c0bd7680
Signed-off-by: Matthew <matthewx.moradi@intel.com>