Update patch set 1
Patch Set 1:

the gate csit has two failures having to do with flows not being
present in the flow table. link here [0], but the exact string it's
not able to find in the flow tables


'arp,arp_op=2 actions=CONTROLLER:65535,resubmit\(,17\)'


that flow is still there, but it's no longer punting to CONTROLLER
and only doing the resubmit(17). Is this a functionality change,
or a bug in the patch? if it's the former, we will need an accompanying
CSIT patch to modify the test.


[0] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-ocata-gate-stateful-carbon/9/log.html.gz#s1-s5-t15-k11

Patch-set: 1