Update patch set 2
Patch Set 2: Code-Review-1

> I see only failures wrt 4 tests related to pipeline changes for
 > GwMAC table that are expected as part of this merge.

Is this in response to my earlier comment/question? are we removing
the punt to CONTROLLER action in the pipeline? That is sort of a 
functionality change. I just want to verify this. If that's the 
case, and you don't have time, I can put the CSIT patch to account
for that. But, we want both CSIT and this patch to be merged around
the same time.

Patch-set: 2
Label: Code-Review=-1