Update patch set 2
Patch Set 2:

> Ran a couple more sandbox jobs. 3 failed with 7 tempest failures,
 > then 4 passed 100%. running a fifth job to see how it looks.
 > 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-shague-upstream-stateful-carbon/4/

Sam, these tempest failures are what we were worried about right?
Currently, with upstream/merged code we are not seeing multiple
tempest failures happening. Every once in a while we will see
a single test fail, but if we are getting multiple in some of these
gate jobs, I think it's something to be careful with.

Patch-set: 2