Update patch set 19
Patch Set 19:

Can you please add code in add() method of vpnInstanceListener where if we see a vpnInstance is being recreated for the same RD list as an anothere vpnInstance which is in pending-delete state, we throw up an ERROR and return.

Doing queued processing of the same can be discussed later.

Patch-set: 19