Bug 8889 - Removed Unused Local Variables
If a local variable is declared but not used, it is dead code and should be removed.
Doing so will improve maintainability because developers will not wonder what the variable is used for.
There were a lot of these that weren't being used according to sonar, so I took the liberty of removing them.

Change-Id: Ica4ad7aa11e91ae6c9ba72ce17af3a479ad1e0be
Signed-off-by: Matthew <matthewx.moradi@intel.com>