Update patch set 3
Patch Set 3:

Question for Vivek: We have likely hit this same issue in other areas - did we solve it the same way to look in the other datastore? Or is there an ordering problem that should be addressed.

Other question is about the null checks. In other recent patches we have been leaning more towards finding the source of the nulls rather than masking them. Does the one her fall into that category or is it OK that the config was deleted already?

Patch-set: 3